Initial compatibility with updated server, retrieves and displays blips on a hardcoded location. Attraction type is customizable, but location is always the same.
Next step is to get device's location and use that instead of the hardcoded location.

Should get rid of custom lookup view, instead replacing it with a "pop-up" mini view (i.e. not whole screen) that allows user
to specify attraction type. Also need an option somewhere to set radius, this can be in an option somewhere? Eventually need to accept
multiple attraction types.